fix: restore backwards compatibility
Note that the current implementation is flawed because custom
ResourceUpdatePreProcessor should be registered to be found when
attempting to use custom matching logic. See also #1914 for more
details.
