+++ operator-framework-core/src/main/java/io/javaoperatorsdk/operator/processing/dependent/kubernetes/processors/GenericResourceUpdatePreProcessor.java
@@ -10,7 +10,6 @@ import io.fabric8.kubernetes.api.model.rbac.ClusterRole;
 import io.fabric8.kubernetes.api.model.rbac.ClusterRoleBinding;
 import io.fabric8.kubernetes.api.model.rbac.Role;
 import io.fabric8.kubernetes.api.model.rbac.RoleBinding;
-import io.fabric8.zjsonpatch.JsonDiff;
 import io.javaoperatorsdk.operator.ReconcilerUtils;
 import io.javaoperatorsdk.operator.api.config.ConfigurationServiceProvider;
 import io.javaoperatorsdk.operator.api.reconciler.Context;
@@ -49,30 +48,5 @@ public class GenericResourceUpdatePreProcessor<R extends HasMetadata> implements
   protected void updateClonedActual(R actual, R desired) {
     var desiredSpec = ReconcilerUtils.getSpec(desired);
     ReconcilerUtils.setSpec(actual, desiredSpec);
-  }
-
-  @Override
-  public boolean matches(R actual, R desired, boolean equality) {
-    final var objectMapper = ConfigurationServiceProvider.instance().getObjectMapper();
-
-    // reflection will be replaced by this:
-    // https://github.com/fabric8io/kubernetes-client/issues/3816
-    var desiredSpecNode = objectMapper.valueToTree(ReconcilerUtils.getSpec(desired));
-    var actualSpecNode = objectMapper.valueToTree(ReconcilerUtils.getSpec(actual));
-    var diffJsonPatch = JsonDiff.asJson(desiredSpecNode, actualSpecNode);
-    // In case of equality is set to true, no diffs are allowed, so we return early if diffs exist
-    // On contrary (if equality is false), "add" is allowed for cases when for some
-    // resources Kubernetes fills-in values into spec.
-    final var diffNumber = diffJsonPatch.size();
-    if (equality && diffNumber > 0) {
-      return false;
-    }
-    for (int i = 0; i < diffNumber; i++) {
-      String operation = diffJsonPatch.get(i).get("op").asText();
-      if (!operation.equals("add")) {
-        return false;
-      }
-    }
-    return true;
   }
 }
